# gpt-docs-proxy-openapi.yaml
openapi: 3.1.0
info:
  title: GPT Google Docs Proxy
  version: 1.0.0
servers:
  - url: https://gpt-docs-proxy.onrender.com
paths:
  /authorize:
    get:
      operationId: authorize
      summary: Redirects to Google OAuth consent screen

  /oauth2callback:
    get:
      operationId: oauth2callback
      summary: Google OAuth callback endpoint

  /docs:
    get:
      operationId: searchDocsByTitle
      summary: Search Google Docs by partial title
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: Title fragment to match
      responses:
        '200':
          description: List of matching docs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string

  /docs/all:
    get:
      operationId: listAllDocs
      summary: List all accessible Google Docs
      responses:
        '200':
          description: List of docs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string

  /docs/{doc_id}:
    get:
      operationId: readDocById
      summary: Read full text of a Google Doc
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
          description: The document’s ID
      responses:
        '200':
          description: Document text
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  text:
                    type: string

  /docs/batch:
    post:
      operationId: batchReadDocs
      summary: Read multiple Google Docs by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                doc_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: List of document texts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    text:
                      type: string